requirements.txt
python==3.8.13
einops==0.4.1
jupyterlab==3.3.2   
h5py==3.6.0
hdf5==1.10.6    
numpy==1.21.2
pandas==1.4.1
pytorch==1.11.0
pyyaml==6.0
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
torchvision==0.12.0
tqdm==4.63.0 
jupyterlab==3.3.2   
matplotlib==3.5.1 

